Adapt to new CONFIG_VIRT default setting
Luckily the QEMU command line in the Makeconf.boot.example for the
QEMU Virt machine already comes with virtualization=true so we
only need to fix it in individual cases where the tutorials
run QEMU manually.

For some reason QEMU does not start bootstrap in HYP mode for
the vexpress-a15 machine, so the user must unselect Virtualization
support in the BUILDING howto.

Change-Id: I9707f3c9caa57d63b580d6f33e83ea7aa6b30048